A few changes to electron density scoring and the density_tools application. - Reduce the number of times rho_calc is computed in density_tools - Mask used by density tools is always bandlimited - Better treatment of mask radus for high B factor atoms - Fix a small bug in scattering factors where they were always integers - Fix incorrect S scattering and add P scattering params
Old repository SHA1: 339943e07bdd1e4d34650247425f2199156ae5dc

+++ source/src/core/scoring/electron_density/xray_scattering.hh
@@ -70,7 +70,7 @@ public:
 		weight_ = 0;
 		sigma_ = 3.0;
 	}
-	OneGaussianScattering(int w, float s) {
+	OneGaussianScattering(float w, float s) {
 		sigma_ = s;
 		weight_ = w;
 	}
@@ -129,7 +129,7 @@ public:
 
 private:
 	float sigma_;
-	int weight_;
+	float weight_;
 };
